- rewritten `_BaseButtonTest` in typescript - refactored/optimized `_BaseButtonTest.tsx` - removed `ButtonTest.jsx` as it turned out to be a set of duplicating tests - minor docs update
