+++ apps/src/componentLibrary/button/_baseButton/_BaseButton.tsx
@@ -297,7 +297,7 @@ const BaseButton: React.FunctionComponent<_BaseButtonProps> = ({
  * * (✔) implementation of component approved by design team;
  * * (✔) has storybook, covered with stories and documentation;
  * * (✔) has tests: test every prop, every state and every interaction that's js related;
- * * (see apps/test/unit/componentLibrary/_BaseButtonTest.jsx)
+ * * (see apps/test/unit/componentLibrary/_BaseButtonTest.tsx)
  * * (?) passes accessibility checks;
  *
  * ###  Status: ```Ready for dev```
@@ -305,6 +305,6 @@ const BaseButton: React.FunctionComponent<_BaseButtonProps> = ({
  * Design System: ***_BaseButton*** Component.
  *
  * ***(!IMPORTANT!)*** This is a private component for Designs System use only. It provides a base for ***Button***
- * and ***LinkButton*** components, implementing all of the logic and styles for them.
+ * and ***LinkButton*** components, implementing all the logic and styles for them.
  */
 export default memo(BaseButton);
